refactor(table): mv to modules
